Revert LATENCY_HDR_SEC_SIGDIGTS to 2. Optimize summarize_current_second by using hdr_value_at_percentiles() instead of N x hdr_value_at_percentile().

+++ run_stats.h
@@ -97,7 +97,7 @@ protected:
     totals m_totals;
 
     std::list<one_second_stats> m_stats;
-    std::vector<float> quantiles_list;
+    std::vector<double> quantiles_list;
 
     // current second stats ( appended to m_stats and reset every second )
     one_second_stats m_cur_stats;
